caduceus: move contentType counter
