+++ src/caduceus/http.go
@@ -98,24 +98,6 @@ func (sh *ServerHandler) ServeHTTP(response http.ResponseWriter, request *http.R
 	debugLog.Log(messageKey, "Request placed on to queue.")
 }
 
-type RequestHandler interface {
-	HandleRequest(workerID int, msg *wrp.Message)
-}
-
-type CaduceusHandler struct {
-	senderWrapper SenderWrapper
-	log.Logger
-}
-
-func (ch *CaduceusHandler) HandleRequest(workerID int, msg *wrp.Message) {
-	msg = fixWrp(msg)
-
-	logging.Info(ch).Log("workerID", workerID, logging.MessageKey(), "Worker received a request, now passing"+
-		" to sender")
-
-	ch.senderWrapper.Queue(msg)
-}
-
 func fixWrp(msg *wrp.Message) *wrp.Message {
 	// "Fix" the WRP if needed.
